# Reduce time taken for backtest.py - def backtest by 60%