Performs bitwise filter on buy/sell rows in the df
before passing to the slower itertuples loop

The larger the dataset the greater the performance gain

Tested for absence of malice with before/after diffs on output of 60 pairs with 24,000 candles. 
No delta on any trades closed or open.  (223 closed trades 19 open)

Diff available here: http://www.mergely.com/EoGF9SST/

Profile traces before and after: 

Before: 4.7 seconds
```
*** PROFILER RESULTS ***
backtest (/Users/creslin/PycharmProjects/freqtrade_new/freqtrade/optimize/backtesting.py:189)
function called 1 times

         3274636 function calls (3261140 primitive calls) in 4.780 seconds
```

After: 2.09 seconds
```
*** PROFILER RESULTS ***
backtest (/Users/creslin/PycharmProjects/freqtrade_new/freqtrade/optimize/backtesting.py:189)
function called 1 times

         1977916 function calls (1961960 primitive calls) in 2.097 seconds


```