@creslinux 2018-07-11
This was working, sadly no SL or ROI were being triggerd. 
Tested with tight Stop/ROI on @xmatthias  suggestion -- lots of delta. 

http://www.mergely.com/0qwuibQH/

Pulling the PR as unsuitable.